+++ consumer/pyproject.toml
@@ -6,6 +6,6 @@ requires-python = ">=3.13"
 dependencies = [
     "aiohttp>=3.12.15",
     "discord-py>=2.6.3",
-    "launchdarkly-eventsource>=1.3.0",
     "psycopg[binary]>=3.2.10",
+    "sans>=1.3.2",
 ]
